New users automatically follow specific libraries
