+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/LibraryCommander.scala
@@ -38,6 +38,7 @@ trait LibraryCommander {
   def modifyLibrary(libraryId: Id[Library], userId: Id[User], modifyReq: LibraryModifications)(implicit context: HeimdalContext): Either[LibraryFail, LibraryModifyResponse]
   def unsafeModifyLibrary(library: Library, modifyReq: LibraryModifications): LibraryModifyResponse
   def deleteLibrary(libraryId: Id[Library], userId: Id[User])(implicit context: HeimdalContext): Option[LibraryFail]
+  def internSystemGeneratedLibraries(userId: Id[User], generateNew: Boolean = true): (Library, Library)
   def createReadItLaterLibrary(userId: Id[User]): Library
   def copyKeepsFromCollectionToLibrary(userId: Id[User], libraryId: Id[Library], tagName: Hashtag)(implicit context: HeimdalContext): Either[LibraryFail, (Seq[Keep], Seq[(Keep, LibraryError)])]
   def moveKeepsFromCollectionToLibrary(userId: Id[User], libraryId: Id[Library], tagName: Hashtag)(implicit context: HeimdalContext): Either[LibraryFail, (Seq[Keep], Seq[(Keep, LibraryError)])]
@@ -455,6 +456,77 @@ class LibraryCommanderImpl @Inject() (
     libraryRepo.save(lib.withOwner(newOwner))
   }
 
+  def internSystemGeneratedLibraries(userId: Id[User], generateNew: Boolean = true): (Library, Library) = {
+    db.readWrite(attempts = 3) { implicit session =>
+      val libMem = libraryMembershipRepo.getWithUserId(userId, None)
+      val allLibs = libraryRepo.getByUser(userId, None)
+      val user = userRepo.get(userId)
+
+      // Get all current system libraries, for main/secret, make sure only one is active.
+      // This corrects any issues with previously created libraries / memberships
+      val sysLibs = allLibs.filter(_._2.ownerId == userId)
+        .filter(l => l._2.kind == LibraryKind.SYSTEM_MAIN || l._2.kind == LibraryKind.SYSTEM_SECRET)
+        .sortBy(_._2.id.get.id)
+        .groupBy(_._2.kind).flatMap {
+          case (kind, libs) =>
+            val (slug, name, visibility) = if (kind == LibraryKind.SYSTEM_MAIN) ("main", "Main Library", LibraryVisibility.DISCOVERABLE) else ("secret", "Secret Library", LibraryVisibility.SECRET)
+
+            if (user.state == UserStates.ACTIVE) {
+              val activeLib = libs.head._2.copy(state = LibraryStates.ACTIVE, slug = LibrarySlug(slug), name = name, visibility = visibility, memberCount = 1)
+              val membership = libMem.find(m => m.libraryId == activeLib.id.get && m.access == LibraryAccess.OWNER)
+              if (membership.isEmpty) airbrake.notify(s"user $userId - non-existing ownership of library kind $kind (id: ${activeLib.id.get})")
+              val activeMembership = membership.getOrElse(LibraryMembership(libraryId = activeLib.id.get, userId = userId, access = LibraryAccess.OWNER)).copy(state = LibraryMembershipStates.ACTIVE)
+              val active = (activeMembership, activeLib)
+              if (libs.tail.length > 0) airbrake.notify(s"user $userId - duplicate active ownership of library kind $kind (ids: ${libs.tail.map(_._2.id.get)})")
+              val otherLibs = libs.tail.map {
+                case (a, l) =>
+                  val inactMem = libMem.find(_.libraryId == l.id.get)
+                    .getOrElse(LibraryMembership(libraryId = activeLib.id.get, userId = userId, access = LibraryAccess.OWNER))
+                    .copy(state = LibraryMembershipStates.INACTIVE)
+                  (inactMem, l.copy(state = LibraryStates.INACTIVE))
+              }
+              active +: otherLibs
+            } else {
+              // do not reactivate libraries / memberships for nonactive users
+              libs
+            }
+        }.toList // force eval
+
+      // save changes for active users only
+      if (sysLibs.nonEmpty && user.state == UserStates.ACTIVE) {
+        sysLibs.map {
+          case (mem, lib) =>
+            libraryRepo.save(lib)
+            libraryMembershipRepo.save(mem)
+        }
+      }
+
+      // If user is missing a system lib, create it
+      val mainOpt = if (!sysLibs.exists(_._2.kind == LibraryKind.SYSTEM_MAIN)) {
+        val mainLib = libraryRepo.save(Library(name = "Main Library", ownerId = userId, visibility = LibraryVisibility.DISCOVERABLE, slug = LibrarySlug("main"), kind = LibraryKind.SYSTEM_MAIN, memberCount = 1, keepCount = 0))
+        libraryMembershipRepo.save(LibraryMembership(libraryId = mainLib.id.get, userId = userId, access = LibraryAccess.OWNER))
+        if (!generateNew) {
+          airbrake.notify(s"$userId missing main library")
+        }
+        searchClient.updateLibraryIndex()
+        Some(mainLib)
+      } else None
+
+      val secretOpt = if (!sysLibs.exists(_._2.kind == LibraryKind.SYSTEM_SECRET)) {
+        val secretLib = libraryRepo.save(Library(name = "Secret Library", ownerId = userId, visibility = LibraryVisibility.SECRET, slug = LibrarySlug("secret"), kind = LibraryKind.SYSTEM_SECRET, memberCount = 1, keepCount = 0))
+        libraryMembershipRepo.save(LibraryMembership(libraryId = secretLib.id.get, userId = userId, access = LibraryAccess.OWNER))
+        if (!generateNew) {
+          airbrake.notify(s"$userId missing secret library")
+        }
+        searchClient.updateLibraryIndex()
+        Some(secretLib)
+      } else None
+
+      val mainLib = sysLibs.find(_._2.kind == LibraryKind.SYSTEM_MAIN).map(_._2).orElse(mainOpt).get
+      val secretLib = sysLibs.find(_._2.kind == LibraryKind.SYSTEM_SECRET).map(_._2).orElse(secretOpt).get
+      (mainLib, secretLib)
+    }
+  }
   def createReadItLaterLibrary(userId: Id[User]): Library = db.readWrite(attempts = 3) { implicit s =>
     val readItLaterLib = libraryRepo.save(Library(name = "Read It Later", ownerId = userId, visibility = LibraryVisibility.SECRET, slug = LibrarySlug("read_id_later"), kind = LibraryKind.SYSTEM_READ_IT_LATER, memberCount = 1, keepCount = 0))
     libraryMembershipRepo.save(LibraryMembership(libraryId = readItLaterLib.id.get, userId = userId, access = LibraryAccess.OWNER))
