+++ kifi-backend/modules/shoebox/test/com/keepit/commanders/KeepInternerTest.scala
@@ -39,7 +39,7 @@ class KeepInternerTest extends Specification with ShoeboxTestInjector {
         val user = db.readWrite { implicit session =>
           UserFactory.user().withName("Shanee", "Smith").withUsername("test").saved
         }
-        val (main, secret) = inject[LibraryInfoCommander].internSystemGeneratedLibraries(user.id.get)
+        val (main, secret) = inject[LibraryCommander].internSystemGeneratedLibraries(user.id.get)
         val bookmarkInterner = inject[KeepInterner]
         val (bookmarks, _) = bookmarkInterner.internRawBookmarks(inject[RawBookmarkFactory].toRawBookmarks(Json.obj(
           "url" -> "http://42go.com",
@@ -59,7 +59,7 @@ class KeepInternerTest extends Specification with ShoeboxTestInjector {
         val user = db.readWrite { implicit session =>
           UserFactory.user().withName("Shanee", "Smith").withUsername("test").saved
         }
-        inject[LibraryInfoCommander].internSystemGeneratedLibraries(user.id.get)
+        inject[LibraryCommander].internSystemGeneratedLibraries(user.id.get)
         val bookmarkInterner = inject[KeepInterner]
         bookmarkInterner.persistRawKeeps(inject[RawKeepFactory].toRawKeep(user.id.get, KeepSource.bookmarkImport, Json.obj(
           "url" -> "http://42go.com",
@@ -80,7 +80,7 @@ class KeepInternerTest extends Specification with ShoeboxTestInjector {
           libraryMembershipRepo.save(LibraryMembership(libraryId = lib.id.get, userId = user.id.get, access = LibraryAccess.OWNER))
           (user, lib)
         }
-        inject[LibraryInfoCommander].internSystemGeneratedLibraries(user.id.get)
+        inject[LibraryCommander].internSystemGeneratedLibraries(user.id.get)
         val bookmarkInterner = inject[KeepInterner]
         bookmarkInterner.persistRawKeeps(inject[RawKeepFactory].toRawKeep(user.id.get, KeepSource.bookmarkImport, Json.obj(
           "url" -> "http://42go.com",
@@ -101,7 +101,7 @@ class KeepInternerTest extends Specification with ShoeboxTestInjector {
         val user = db.readWrite { implicit session =>
           UserFactory.user().withName("Shanee", "Smith").withUsername("test").saved
         }
-        val (library, _) = inject[LibraryInfoCommander].internSystemGeneratedLibraries(user.id.get)
+        val (library, _) = inject[LibraryCommander].internSystemGeneratedLibraries(user.id.get)
         val bookmarkInterner = inject[KeepInterner]
         val raw = inject[RawBookmarkFactory].toRawBookmarks(Json.arr(Json.obj(
           "url" -> "http://42go.com",
@@ -146,7 +146,7 @@ class KeepInternerTest extends Specification with ShoeboxTestInjector {
         val user = db.readWrite { implicit session =>
           UserFactory.user().withName("Shanee", "Smith").withUsername("test").saved
         }
-        val (library, _) = inject[LibraryInfoCommander].internSystemGeneratedLibraries(user.id.get)
+        val (library, _) = inject[LibraryCommander].internSystemGeneratedLibraries(user.id.get)
         val fakeAirbrake = inject[FakeAirbrakeNotifier]
         fakeAirbrake.errorCount() === 0
         val bookmarkInterner = inject[KeepInterner]
@@ -179,7 +179,7 @@ class KeepInternerTest extends Specification with ShoeboxTestInjector {
         val user = db.readWrite { implicit s =>
           UserFactory.user().withName("Greg", "Smith").withUsername("test").saved
         }
-        val (library, _) = inject[LibraryInfoCommander].internSystemGeneratedLibraries(user.id.get)
+        val (library, _) = inject[LibraryCommander].internSystemGeneratedLibraries(user.id.get)
         val bookmarkInterner = inject[KeepInterner]
         val (initialBookmarks, _) = bookmarkInterner.internRawBookmarks(inject[RawBookmarkFactory].toRawBookmarks(Json.arr(Json.obj(
           "url" -> "http://42go.com/",
