+++ kifi-backend/modules/shoebox/test/com/keepit/commanders/RawKeepImporterTest.scala
@@ -61,7 +61,7 @@ class RawKeepImporterTest extends TestKitSupport with SpecificationLike with Sho
 
           (u, l)
         }
-        inject[LibraryInfoCommander].internSystemGeneratedLibraries(user.id.get)
+        inject[LibraryCommander].internSystemGeneratedLibraries(user.id.get)
         val bookmarkInterner = inject[KeepInterner]
         val json = Json.parse(io.Source.fromFile(new File("test/data/bookmarks_small.json"), UTF8).mkString)
         bookmarkInterner.persistRawKeeps(inject[RawKeepFactory].toRawKeep(user.id.get, KeepSource.bookmarkImport, json, libraryId = lib.id))
@@ -90,7 +90,7 @@ class RawKeepImporterTest extends TestKitSupport with SpecificationLike with Sho
           libraryMembershipRepo.save(LibraryMembership(libraryId = lib.id.get, userId = user.id.get, access = LibraryAccess.OWNER))
           (user, lib)
         }
-        inject[LibraryInfoCommander].internSystemGeneratedLibraries(user.id.get)
+        inject[LibraryCommander].internSystemGeneratedLibraries(user.id.get)
         val bookmarkInterner = inject[KeepInterner]
         val json = Json.parse(io.Source.fromFile(new File("test/data/bookmarks_small.json"), UTF8).mkString)
         bookmarkInterner.persistRawKeeps(inject[RawKeepFactory].toRawKeep(user.id.get, KeepSource.bookmarkImport, json, libraryId = Some(lib.id.get)))
