+++ kifi-backend/modules/shoebox/test/com/keepit/controllers/mobile/MobileKeepsControllerTest.scala
@@ -681,7 +681,7 @@ class MobileKeepsControllerTest extends Specification with ShoeboxTestInjector w
         val user = db.readWrite { implicit session =>
           UserFactory.user().withName("Eishay", "Smith").withUsername("test").saved
         }
-        inject[LibraryInfoCommander].internSystemGeneratedLibraries(user.id.get)
+        inject[LibraryCommander].internSystemGeneratedLibraries(user.id.get)
 
         val withCollection =
           RawBookmarkRepresentation(title = Some("title 11"), url = "http://www.hi.com11", isPrivate = None) ::
@@ -731,7 +731,7 @@ class MobileKeepsControllerTest extends Specification with ShoeboxTestInjector w
           UserFactory.user().withName("Eishay", "Smith").withUsername("test").saved
         }
 
-        inject[LibraryInfoCommander].internSystemGeneratedLibraries(user.id.get)
+        inject[LibraryCommander].internSystemGeneratedLibraries(user.id.get)
 
         val keep1ToCollections = (Json.obj("title" -> "title 11", "url" -> "http://www.hi.com11", "isPrivate" -> false), Seq("tagA", "tagB", "tagC"))
         val keep2ToCollections = (Json.obj("title" -> "title 11", "url" -> "http://www.hi.com11", "isPrivate" -> false), Seq("tagA", "tagD", "tagE"))
